♻️ refactor: align AgentRegistration with codebase conventions
- Extract all styled components to styled.ts (Hans/codebase pattern)
- Extract AGENT_TYPE_LABELS and validateStep to helpers.ts (removes duplication)
- Add apiPathUser to constants.ts (was locally defined in component)
- Switch interface Props → type Props across all step components
- Replace hardcoded hex error colours with CSS variable fallbacks
